requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.1
boto3==1.24.53
botocore==1.27.53
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Django==3.2.19
Flask==2.2.2
fonttools==4.35.0
humanize==4.6.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
packaging==21.3
pandas==1.3.5
Pillow==9.5.0
psutil==5.9.5
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
tabulate==0.9.0
tinydb==4.5.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.15.0
zipp==3.8.1
